MDSS from CM 3.4 kernel
+ PWM changes from LA.BR branch + some not mandatory stability changes

+++ drivers/video/msm/mdss/mdss_qpic.c
@@ -1,4 +1,4 @@
-/* Copyright (c) 2013-2014, The Linux Foundation. All rights reserved.
+/* Copyright (c) 2013, The Linux Foundation. All rights reserved.
  *
  * This program is free software; you can redistribute it and/or modify
  * it under the terms of the GNU General Public License version 2 and
@@ -28,9 +28,11 @@
 #include <linux/uaccess.h>
 #include <linux/bootmem.h>
 #include <linux/dma-mapping.h>
-#include <linux/clk/msm-clk.h>
 
-#include <linux/msm-sps.h>
+#include <asm/system.h>
+#include <asm/mach-types.h>
+#include <mach/sps.h>
+#include <linux/clk/msm-clk.h>
 #include <mach/hardware.h>
 
 #include "mdss_fb.h"
@@ -201,7 +203,7 @@ int qpic_init_sps(struct platform_device *pdev,
 	struct sps_connect *sps_config = &end_point->config;
 	struct sps_register_event *sps_event = &end_point->bam_event;
 	struct sps_bam_props bam = {0};
-	unsigned long bam_handle = 0;
+	u32 bam_handle = 0;
 
 	if (qpic_res->sps_init)
 		return 0;
